refactor: extract rpc module
